src/prompt-box.vala: enable line wrapping for (error) messages.
Long lines were previously truncated as part of the fading label. Since
PAM (error) messages are shown within the prompt box, truncating them is
a really bad idea, since the user might not understand what went wrong
and how it can be fixed.

Enabling line wrapping makes sure that the full message is always shown
- at least theoretically. Messages may still be truncated if they are
long enough to overflow the prompt box itself, but this is caused by the
prompt box not resizing itself correctly. Further work on that is
necessary.