src/util.vala: add contrast_color helper, generating a contrast color for a given color.
A contrast color is usually either black or white - whatever enhances
contrast to the given input color.

The reason for this is complicated and best explained in the referred to
material in the comments, but it generally boils down to these two
colors (or lack of color or a mixture of all colors, whatever you
prefer) creating the best contrast from a human perceptive viewpoint.